moved scrolling logic from Conversation to ConversationBody
