+++ client/src/ui/messaging/Conversation.tsx
@@ -4,7 +4,7 @@ import { ConversationHeader } from "./ConversationHeader";
 import { ConversationBody } from "./ConversationBody";
 import Image from "next/image";
 import { useMessage } from "../../lib/store/message";
-import { useEffect, useState, useRef } from "react";
+import { useState, useRef } from "react";
 import { useProfile } from "../../lib/store/profile";
 import { useSubscribeConversation } from "@/lib/hooks/useSubscribeConversation";
 
@@ -23,24 +23,12 @@ export function Conversation({ conversationId }: { conversationId: string }) {
     setNumNewMessages(0);
   }
 
-  // Scroll to the bottom of the element
-  function scrollDown() {
-    if (bottomRef.current) {
-      bottomRef.current.scrollIntoView({ behavior: "smooth" });
-    }
-  }
-
   useSubscribeConversation(
     conversationId,
     addMessage,
     profile,
     setNumNewMessages
   );
-
-  // scrolls down to the latest message on page mount
-  useEffect(() => {
-    scrollDown();
-  }, []); // Empty dependency array ensures this runs only once on mount
 
   return (
     <div
@@ -71,7 +59,6 @@ export function Conversation({ conversationId }: { conversationId: string }) {
         messages={messages}
         numNewMessages={numNewMessages}
         resetNumNewMessages={() => resetNumNewMessages()}
-        scrollDown={scrollDown}
         bottomRef={bottomRef}
       />
